requirements.txt
playwright
requests
python-telegram-bot
python-dotenv
gspread
google-auth